Added support for trim-operators and TRIMRANGE function.
