+++ src/EPPlus/FormulaParsing/Excel/Functions/RefAndLookup/TrimFunctions/TrimRows.cs
@@ -1,0 +1,25 @@
+﻿/*************************************************************************************************
+  Required Notice: Copyright (C) EPPlus Software AB. 
+  This software is licensed under PolyForm Noncommercial License 1.0.0 
+  and may only be used for noncommercial purposes 
+  https://polyformproject.org/licenses/noncommercial/1.0.0/
+
+  A commercial license to use this software can be purchased at https://epplussoftware.com
+ *************************************************************************************************
+  Date               Author                       Change
+ *************************************************************************************************
+  19/3/2026         EPPlus Software AB           EPPlus v8.6
+ *************************************************************************************************/
+namespace OfficeOpenXml.FormulaParsing.Excel.Functions.RefAndLookup.TrimFunctions
+{
+    internal partial class TrimRange
+    {
+        private enum TrimRows
+        {
+            None = 0,
+            TrimLeading = 1,
+            TrimTrailing = 2,
+            TrimAll = 3
+        }
+    }
+}
