+++ src/EPPlus/FormulaParsing/Excel/Functions/RefAndLookup/TrimFunctions/TroTrailing.cs
